Move WEBSITE_CATEGORY_MAP to data-constants.ts
Separate the domain-to-category mapping into its own file to keep
app-constants focused on IPC/app-level constants and data-constants
for data mappings.
