+++ src/constants/app-constants.ts
@@ -283,212 +283,6 @@ export const WEBSITE_CATEGORY_COLORS: Record<string, string> = {
 };
 
 /**
- * Domain → website category mapping.
- * Used by the bookmarks page, history page, and other pages to auto-categorise websites.
- */
-export const WEBSITE_CATEGORY_MAP: Record<string, string> = {
-  // Dev
-  'github.com': 'dev',
-  'gitlab.com': 'dev',
-  'bitbucket.org': 'dev',
-  'stackoverflow.com': 'dev',
-  'stackexchange.com': 'dev',
-  'developer.mozilla.org': 'dev',
-  'electronjs.org': 'dev',
-  'nodejs.org': 'dev',
-  'npmjs.com': 'dev',
-  'pypi.org': 'dev',
-  'crates.io': 'dev',
-  'pkg.go.dev': 'dev',
-  'vitepress.dev': 'dev',
-  'vitejs.dev': 'dev',
-  'source.chromium.org': 'dev',
-  'caniuse.com': 'dev',
-  'css-tricks.com': 'dev',
-  'web.dev': 'dev',
-  'dev.to': 'dev',
-  'medium.com': 'dev',
-  'hashnode.dev': 'dev',
-  'codepen.io': 'dev',
-  'jsfiddle.net': 'dev',
-  'replit.com': 'dev',
-  'logrocket.com': 'dev',
-  'smashingmagazine.com': 'dev',
-  'freecodecamp.org': 'dev',
-  'typescriptlang.org': 'dev',
-  'rust-lang.org': 'dev',
-  'go.dev': 'dev',
-  'docs.python.org': 'dev',
-  'reactjs.org': 'dev',
-  'vuejs.org': 'dev',
-  'svelte.dev': 'dev',
-  'angular.io': 'dev',
-  'webpack.js.org': 'dev',
-  'vercel.com': 'dev',
-  'netlify.com': 'dev',
-  'heroku.com': 'dev',
-  'docker.com': 'dev',
-  'kubernetes.io': 'dev',
-  'linear.app': 'dev',
-  'jira.atlassian.com': 'dev',
-
-  // News
-  'news.ycombinator.com': 'news',
-  'arstechnica.com': 'news',
-  'theverge.com': 'news',
-  'techcrunch.com': 'news',
-  'wired.com': 'news',
-  'bbc.com': 'news',
-  'reuters.com': 'news',
-  'nytimes.com': 'news',
-  'theguardian.com': 'news',
-  'washingtonpost.com': 'news',
-  'apnews.com': 'news',
-  'cnn.com': 'news',
-  'bloomberg.com': 'news',
-  'engadget.com': 'news',
-  'zdnet.com': 'news',
-  'theregister.com': 'news',
-  'privacyguides.org': 'news',
-  'eff.org': 'news',
-  'torrentfreak.com': 'news',
-
-  // Media
-  'youtube.com': 'media',
-  'vimeo.com': 'media',
-  'spotify.com': 'media',
-  'soundcloud.com': 'media',
-  'twitch.tv': 'gaming',
-  'netflix.com': 'media',
-  'disneyplus.com': 'media',
-  'podcasts.apple.com': 'media',
-  'music.apple.com': 'media',
-  'bandcamp.com': 'media',
-  'dailymotion.com': 'media',
-  'imgur.com': 'media',
-  'flickr.com': 'media',
-  'unsplash.com': 'media',
-
-  // Social
-  'reddit.com': 'social',
-  'twitter.com': 'social',
-  'x.com': 'social',
-  'mastodon.social': 'social',
-  'linkedin.com': 'social',
-  'facebook.com': 'social',
-  'instagram.com': 'social',
-  'threads.net': 'social',
-  'discord.com': 'social',
-  'slack.com': 'social',
-  'lobste.rs': 'social',
-
-  // Productivity
-  'notion.so': 'productivity',
-  'trello.com': 'productivity',
-  'docs.google.com': 'productivity',
-  'drive.google.com': 'productivity',
-  'sheets.google.com': 'productivity',
-  'gmail.com': 'productivity',
-  'outlook.com': 'productivity',
-  'calendar.google.com': 'productivity',
-
-  // Tools
-  'translate.google.com': 'tools',
-  'excalidraw.com': 'tools',
-  'regex101.com': 'tools',
-  'jsonformatter.org': 'tools',
-
-  // Finance
-  'finance.yahoo.com': 'finance',
-  'coinmarketcap.com': 'finance',
-  'tradingview.com': 'finance',
-
-  // Shopping
-  'amazon.com': 'shopping',
-  'ebay.com': 'shopping',
-  'etsy.com': 'shopping',
-
-  // Reference
-  'wikipedia.org': 'reference',
-  'en.wikipedia.org': 'reference',
-  'wikimedia.org': 'reference',
-  'archive.org': 'reference',
-  'wolframalpha.com': 'reference',
-  'arxiv.org': 'reference',
-  'mdn.io': 'reference',
-
-  // Search
-  'google.com': 'search',
-  'bing.com': 'search',
-  'duckduckgo.com': 'search',
-
-  // Design
-  'figma.com': 'design',
-  'canva.com': 'design',
-  'dribbble.com': 'design',
-  'behance.net': 'design',
-
-  // Health
-  'webmd.com': 'health',
-  'mayoclinic.org': 'health',
-  'cdc.gov': 'health',
-  'nih.gov': 'health',
-  'healthline.com': 'health',
-  'medlineplus.gov': 'health',
-  'myfitnesspal.com': 'health',
-  'strava.com': 'health',
-  'fitbit.com': 'health',
-
-  // Gaming
-  'store.steampowered.com': 'gaming',
-  'steampowered.com': 'gaming',
-  'ign.com': 'gaming',
-  'roblox.com': 'gaming',
-  'epicgames.com': 'gaming',
-  'gog.com': 'gaming',
-  'kotaku.com': 'gaming',
-  'gamespot.com': 'gaming',
-  'polygon.com': 'gaming',
-
-  // Travel
-  'booking.com': 'travel',
-  'airbnb.com': 'travel',
-  'expedia.com': 'travel',
-  'tripadvisor.com': 'travel',
-  'kayak.com': 'travel',
-  'skyscanner.com': 'travel',
-  'hotels.com': 'travel',
-  'google.com/travel': 'travel',
-
-  // Education
-  'coursera.org': 'education',
-  'khanacademy.org': 'education',
-  'udemy.com': 'education',
-  'edx.org': 'education',
-  'udacity.com': 'education',
-  'skillshare.com': 'education',
-  'leetcode.com': 'education',
-  'duolingo.com': 'education',
-  'brilliant.org': 'education',
-
-  // Entertainment
-  'espn.com': 'entertainment',
-  'yelp.com': 'entertainment',
-  'doordash.com': 'entertainment',
-  'ubereats.com': 'entertainment',
-  'grubhub.com': 'entertainment',
-  'allrecipes.com': 'entertainment',
-  'imdb.com': 'entertainment',
-  'rottentomatoes.com': 'entertainment',
-  'buzzfeed.com': 'entertainment',
-  'tinder.com': 'entertainment',
-  'bumble.com': 'entertainment',
-  'cars.com': 'entertainment',
-  'autotrader.com': 'entertainment',
-  'zillow.com': 'entertainment',
-};
-
-/**
  * Known multi-part country-code TLDs for correct eTLD+1 extraction.
  * Used by cookie third-party detection to avoid misclassifying subdomains.
  */
