+++ src/renderer/pages/history/index.ts
@@ -1,7 +1,8 @@
 import { HtmlUtils } from '../../../renderer/common/html-utils';
 import { FormatUtils } from '../../../renderer/common/format-utils';
 import { BrowsingHistoryRecord } from '../../../types/browsing-history-record';
-import { WEBSITE_CATEGORY_MAP, WEBSITE_CATEGORY_COLORS } from '../../../constants/app-constants';
+import { WEBSITE_CATEGORY_COLORS } from '../../../constants/app-constants';
+import { WEBSITE_CATEGORY_MAP } from '../../../constants/data-constants';
 import './index.css';
 
 import { createIcons, icons } from 'lucide';
